refactor: move TigrblApi to tigrbl.api
